Added MenuBuilderExample to check the code like they asked is section c int the exercise
