+++ MenuBuilder/MenuBuilder-Example/IntMenu/GoToMenuAction.cs
@@ -1,8 +1,9 @@
-﻿using System;
+﻿using MenuBuilder;
+using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace MenuBuilder.IntMenu
+namespace MenuBuilder_Example.IntMenu
 {
     public class GoToMenuAction : IActions<int>
     {
