+++ MenuBuilder/MenuBuilder-Example/IntMenu/IntegersMenu.cs
@@ -1,28 +1,30 @@
-﻿using System;
+﻿using MenuBuilder;
+using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
-using System.Security.Cryptography;
 using System.Text;
 
-namespace MenuBuilder.StringMenu
+namespace MenuBuilder_Example.IntMenu
 {
-    public class StringMenu : BasicMenu<string>
+    public class IntegersMenu : BasicMenu<int>
     {
-        private const string ENTER_VARIABLES = "Enter variables";
         private const string ERROR_MSG = "Invalid input";
-        private readonly List<string> _requaiersInput = new List<string>();
+        private const string ENTER_VARIABLES = "Enter variables";
+        private readonly List<string> _requaiersInput = new List<string> { "1", "2" };
 
-        private Dictionary<string, IActions<string>> _options;
+        private Dictionary<string, IActions<int>> _options;
 
         private Dictionary<string, string> _displayOptions;
 
-        public StringMenu()
+        public IntegersMenu()
         {
+            _options = new Dictionary<string, IActions<int>>();
             _displayOptions = new Dictionary<string, string>();
-            _options = new Dictionary<string, IActions<string>>();
+            
         }
 
-        public override void AddAction(string option, string description, IActions<string> action, bool requiersInput=false)
+        public override void AddAction(string option, string description, IActions<int> action, bool requiersInput=false)
         {
             _options.Add(option, action);
             _displayOptions.Add(option, description);
@@ -34,21 +36,21 @@ namespace MenuBuilder.StringMenu
 
         public override void Main()
         {
-            BasicValidator validator = new BasicValidator(_options.Keys.ToList());
             ConsoleDisplayer consoleDisplayer = new ConsoleDisplayer(_displayOptions);
-            string input;
+            IntValidator validator = new IntValidator(_options.Keys.ToList());
+            string input = string.Empty;
             while (true)
             {
                 consoleDisplayer.ShowOptions();
                 input = Console.ReadLine();
                 bool isInputValid = validator.Validate(input);
-                if (isInputValid)
+                if(isInputValid)
                 {
                     if (_requaiersInput.Contains(input))
                     {
                         consoleDisplayer.PrintValueToConsole(ENTER_VARIABLES);
                         string userInput = Console.ReadLine();
-                        string[] variables = userInput.Split(',');
+                        int[] variables = userInput.Split(',').Select(v => Int32.Parse(v)).ToArray();
                         RunOption(input, variables);
                     }
                     else
@@ -63,7 +65,7 @@ namespace MenuBuilder.StringMenu
             }
         }
 
-        public override void RunOption(string option, params string[] variables)
+        public override void RunOption(string option, params int[] variables)
         {
             _options[option].Action(variables);
         }
