+++ MenuBuilder/MenuBuilder-Example/MenuBuilder-Example.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <RootNamespace>MenuBuilder_Example</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\MenuBuilder\MenuBuilder.csproj" />
+  </ItemGroup>
+
+</Project>
