+++ MenuBuilder/MenuBuilder.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30225.117
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MenuBuilder", "MenuBuilder\MenuBuilder.csproj", "{A1403067-9DAF-4874-BD01-368351B73853}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MenuBuilder", "MenuBuilder\MenuBuilder.csproj", "{A1403067-9DAF-4874-BD01-368351B73853}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MenuBuilder-Example", "MenuBuilder-Example\MenuBuilder-Example.csproj", "{50A1169C-EFFC-4405-9357-E13FBFE1961C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{A1403067-9DAF-4874-BD01-368351B73853}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A1403067-9DAF-4874-BD01-368351B73853}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A1403067-9DAF-4874-BD01-368351B73853}.Release|Any CPU.Build.0 = Release|Any CPU
+		{50A1169C-EFFC-4405-9357-E13FBFE1961C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{50A1169C-EFFC-4405-9357-E13FBFE1961C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{50A1169C-EFFC-4405-9357-E13FBFE1961C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{50A1169C-EFFC-4405-9357-E13FBFE1961C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
